Updated db entity and wiring for new column types. Related to Issue [WTI-986]
git-svn-id: https://repos.deltares.nl/repos/WettelijkToetsInstrumentarium/trunk@7315 5a3db67b-de53-47b5-99c8-a1c30a6650e2

Former-commit-id: 6774b9c88da241cdc30ba4bd74e72a315511b65a
Former-commit-id: 2b72d9624adc3799e29d94acbe3b73095198c147

+++ Application/Ringtoets/src/Application.Ringtoets.Storage/Create/AssessmentSectionCreateExtensions.cs
@@ -69,7 +69,7 @@ namespace Application.Ringtoets.Storage.Create
             {
                 Id = section.Id.DeepClone(),
                 Name = section.Name.DeepClone(),
-                Composition = (short) section.Composition,
+                Composition = (byte) section.Composition,
                 Comments = section.Comments.Body.DeepClone(),
                 Norm = section.FailureMechanismContribution.Norm,
                 Order = order
