+++ Application/Ringtoets/src/Application.Ringtoets.Storage/DbContext/CharacteristicPointEntity.cs
@@ -37,7 +37,7 @@ namespace Application.Ringtoets.Storage.DbContext
     {
         public long CharacteristicPointEntityId { get; set; }
         public long SurfaceLineEntityId { get; set; }
-        public short Type { get; set; }
+        public byte Type { get; set; }
         public Nullable<double> X { get; set; }
         public Nullable<double> Y { get; set; }
         public Nullable<double> Z { get; set; }
